requirements.txt
# --- Dependencias principais do backend FastAPI ---
fastapi>=0.111.0           # Framework web
uvicorn[standard]>=0.30.0  # ASGI server para FastAPI

# --- Validacao / settings ---
pydantic-settings>=2.2.1   # Carrega variaveis do .env

# --- RAG / IA ---
sentence-transformers>=2.7.0  # Embeddings
openai>=1.25.0                # Cliente OpenAI
chromadb>=0.4.24              # Vector Store persistente
langchain>=0.1.11             # RecursiveCharacterTextSplitter

# --- Uploads e arquivos ---
aiofiles>=23.2.1          # Leitura assincrona de arquivos
python-multipart>=0.0.9   # Form-data em FastAPI

# --- Interface web opcional ---
streamlit>=1.34.0         # Dashboard simples

# --- Utilidades ---
python-dotenv>=1.0.1      # Carrega .env em scripts standalone
typing_extensions>=4.11.0 # Tipagem retrocompativel

# --- Dependencias de desenvolvimento/teste ---
pytest>=8.0.0,<9.0.0
pytest-cov>=4.0.0,<6.0.0 # Opcional: para cobertura de testes
requests>=2.31.0,<2.33.0 # TestClient do FastAPI usa requests
numpy>=1.26.0,<2.0.0
pytest-asyncio>=0.23.0,<0.24.0
